requirements.txt
accelerate==0.28.0
aiohttp==3.8.6
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.7.0
anthropic==0.3.11
antlr4-python3-runtime==4.9.3
anyio==3.5.0
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
av==10.0.0
azure-cognitiveservices-speech==1.37.0
backoff==2.2.1
bitsandbytes==0.43.0
black==24.4.2
blinker==1.6.3
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.2
cohere==4.32
comm==0.1.4
contextlib2==21.6.0
courlan==0.9.4
crashtest==0.4.1
cryptography==41.0.5
dateparser==1.2.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docker==6.1.3
docstring_parser==0.16
dulwich==0.21.7
einops==0.7.0
elevenlabs==1.2.2
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.3
fastavro==1.8.2
fastjsonschema==2.20.0
filelock==3.15.4
flash-attn==2.5.6
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
google-crc32c==1.5.0
greenlet==3.0.3
griptape==0.24.2
h11==0.14.0
htmldate==1.5.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
ifaddr==0.2.0
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
interegular==0.3.3
ipykernel==6.26.0
ipython==8.17.2
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.5.0
jupyter_core==5.5.0
jusText==3.0.0
keyring==24.3.1
langcodes==3.3.0
lark==1.1.9
llvmlite==0.42.0
lm-format-enforcer==0.9.8
loguru==0.7.2
lxml==4.9.3
mail-parser==3.15.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.3.0
motor==3.3.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.3
ninja==1.11.1.1
numexpr==2.10.0
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.550.52
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.35.13
opencv-python==4.9.0.80
opensearch-py==2.3.2
orjson==3.10.3
oscrypto==1.3.0
outlines==0.0.34
packaging==24.1
pandas==2.2.2
parso==0.8.3
pathspec==0.11.2
peft==0.9.0
pexpect==4.8.0
pgvector==0.2.3
Pillow==10.1.0
pinecone-client==2.2.4
pkginfo==1.11.1
platformdirs==3.11.0
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-redis==1.1.1
pyarrow==13.0.0
pyarrow-hotfix==0.6
PyAudio==0.2.13
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.8.1b0
pydub==0.25.1
pyee==11.0.1
Pygments==2.18.0
PyJWT==2.8.0
pylibsrtp==0.8.0
pymongo==4.8.0
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyOpenSSL==23.3.0
PyPDF2==3.0.1
pyproject_hooks==1.1.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.9.4
ray==2.21.0
redis==5.0.7
referencing==0.30.2
regex==2024.5.15
requests==2.32.3
requests-aws4auth==1.2.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.10.6
ruff==0.4.4
safetensors==0.4.3
schema==0.7.7
scipy==1.13.0
SecretStorage==3.3.3
sentencepiece==0.1.99
shellingham==1.5.4
shtab==1.7.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.52
sqlalchemy-redshift==0.8.14
sse-starlette==1.6.5
stack-data==0.6.3
starlette==0.37.2
stopit==1.1.2
stringcase==1.2.0
sympy==1.13.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
timeit-decorator==1.1.1
tld==0.13
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.3.0
tornado==6.3.3
tqdm==4.66.4
trafilatura==1.6.2
traitlets==5.13.0
transformers==4.40.2
triton==2.3.0
trl==0.8.0
trove-classifiers==2024.7.2
typer==0.12.3
typing_extensions==4.12.2
tyro==0.7.3
tzdata==2023.3
tzlocal==5.2
ujson==5.9.0
urllib3==2.0.3
uvicorn==0.23.2
uvloop==0.19.0
validators==0.22.0
virtualenv==20.26.3
vllm-nccl-cu12==2.18.1.0.4.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.9
websocket-client==1.8.0
websockets==12.0
xformers==0.0.26.post1
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
